fix: Coverage Area, Location UI, and Vendor Card Navigate Function
