+++ src/app/product/[productId]/container/ProductDetailContainer.tsx
@@ -8,6 +8,7 @@ import { MessageIcon } from '@/shared/container/Icon/MessageIcon';
 import { SwiperContainer } from '@/shared/container/Swiper/SwiperContainer';
 import { SwiperSlide } from 'swiper/react';
 import { VendorCard } from '@/shared/container/Card/VendorCard';
+import CoverageAreaSection from './section/CoverageAreaSection';
 import DetailFooter from '@/shared/container/DetailFooter/DetailFooter';
 import DetailHeader from '@/shared/container/DetailHeader/DetailHeader';
 import DetailInfoSection from '@/shared/container/Section/DetailInfoSection';
@@ -54,7 +55,6 @@ const ProductDetailContainer = ({
           sold={20}
           totalReview={12}
           rating={product.rating}
-          coverage_area={product.coverage_area}
         />
         <section>
           <SwiperContainer>
@@ -63,7 +63,8 @@ const ProductDetailContainer = ({
                 key={index}
                 className={`w-[180px] h-[135px] ${index === 0 && 'ml-4'} ${
                   index + 1 === product.images.length && 'mr-4'
-                }`}>
+                }`}
+              >
                 <Image
                   src={image}
                   alt={product.title}
@@ -74,6 +75,7 @@ const ProductDetailContainer = ({
             ))}
           </SwiperContainer>
         </section>
+        <CoverageAreaSection coverage_area={product.coverage_area} />
         <section className="space-y-3 px-4">
           <h3 className="text-body-2 font-medium">Description</h3>
           <p className="text-caption-1 text-ny-gray-400">
@@ -83,8 +85,8 @@ const ProductDetailContainer = ({
         <section className="space-y-3 px-4">
           <h2 className="text-body-2 font-medium">Vendor</h2>
           <VendorCard
-            navigateTo="/"
             id={product.vendor.id}
+            navigateTo={`/vendor/${product.vendor.id}`}
             vendor_name={product.vendor.name}
             product_type_name={product.vendor.type_name}
             price={product.vendor.lowest_price}
@@ -104,12 +106,14 @@ const ProductDetailContainer = ({
         <div className="flex items-center gap-2">
           <Button
             icon={<MessageIcon />}
-            className="flex items-center justify-center flex-1 rounded-[8px] h-[40px] bg-ny-primary-100 text-ny-primary-500 text-body-2">
+            className="flex items-center justify-center flex-1 rounded-[8px] h-[40px] bg-ny-primary-100 text-ny-primary-500 text-body-2"
+          >
             Message
           </Button>
           <Button
             icon={<CartIcon />}
-            className="flex items-center justify-center flex-1 rounded-[8px] h-[40px] bg-ny-primary-500 text-white text-body-2">
+            className="flex items-center justify-center flex-1 rounded-[8px] h-[40px] bg-ny-primary-500 text-white text-body-2"
+          >
             Add to Cart
           </Button>
         </div>
